# Optimize volume_montecarlo performance