@pannarale 2025-05-14
This does not affect PyGRB, as far as I know.