@tdent 2025-05-14
What specific example function call was made to obtain the advertised speedups?